Prevent some warnings in case CrossWorks for ARM is used.
git-svn-id: svn+ssh://svn.code.sf.net/p/ethernut/code/trunk@5500 1ca430b3-8271-4a6b-9b0e-3dcac98a8924

+++ nut/pro/uhttp/modules/mod_ssi.c
@@ -203,7 +203,7 @@ int HttpSsiParse(HTTPD_SESSION *hs, const char *buf, int len)
         buf++;
         len--;
     }
-    for (i = 0; i < HTTP_SSI_NUM_COMMANDS; i++) {
+    for (i = 0; i < (int)HTTP_SSI_NUM_COMMANDS; i++) {
         if (len > ssiCmdList[i].icmd_namelen &&
             strncasecmp(buf, ssiCmdList[i].icmd_name, ssiCmdList[i].icmd_namelen) == 0) {
             ssiCmd = &ssiCmdList[i];
